requirements.txt
asgiref==3.2.7
backcall==0.2.0
beautifulsoup4==4.9.1
boto3==1.14.23
botocore==1.17.23
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap3==12.1.0
django-bootstrap4==2.0.1
django-countries==6.1.2
django-crispy-forms==1.9.1
django-db-log==2.2.1
django-debug-toolbar==2.2
django-extensions==3.0.4
django-filter==2.3.0
django-heroku==0.3.1
django-resized==0.3.11
django-stdimage==5.1.1
django-storages==1.9.1
django-tables2==2.3.1
django-tagging==0.5.0
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.8
docutils==0.15.2
factory-boy==2.12.0
Faker==4.1.1
gunicorn==20.0.4
idna==2.9
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
parso==0.7.1
pickleshare==0.7.5
Pillow==7.1.2
progressbar2==3.51.4
prompt-toolkit==3.0.5
psycopg2==2.8.5
pydotplus==2.0.2
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.48.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.1.0
